server/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
mongoengine==0.27.0
python-dotenv==1.0.0
reportlab==4.0.0
dnspython==2.6.1
